requirements.txt
aiohttp==3.12.13
boto3==1.38.11
cryptography==45.0.6
face_recognition==1.3.0
fastapi==0.116.1
insightface==0.7.3
numpy==2.3.2
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
passlib==1.7.4
Pillow==11.3.0
prometheus_client==0.22.1
prometheus_fastapi_instrumentator==7.1.0
pydantic==2.11.7
pydantic_settings==2.10.1
PyJWT==2.10.1
scikit_learn==1.7.1
SQLAlchemy==2.0.43
starlette==0.47.2
uvicorn==0.35.0
